requirements.txt
Flask==1.1.1
gunicorn==19.9.0
pyrebase4